Major refactoring in progress. Porting to Spring JPA.

+++ src/main/java/se/melsom/warehouse/report/component/Table.java
@@ -1,16 +1,7 @@
 package se.melsom.warehouse.report.component;
 
 
-/**
- * The type Table.
- */
 public class Table extends Frame {
-    /**
-     * Instantiates a new Table.
-     *
-     * @param x the x
-     * @param y the y
-     */
     public Table(float x, float y) {
 		super(x, y, 0, 0);
 	}
@@ -20,7 +11,7 @@ public class Table extends Frame {
 		float width = 0;
 		
 		for (Component row : getComponents()) {
-			width = (float) Math.max(width, row.getWidth());
+			width = Math.max(width, row.getWidth());
 		}
 		
 		return width;
